+++ infra/main.tf
@@ -1,4 +1,6 @@
-#!terraform
+# ---------------------------------------------------------------------------------------------
+# Terraform Configuration ---------------------------------------------------------------------
+# ---------------------------------------------------------------------------------------------
 
 terraform {
   required_version = "~> 1.10.5"
@@ -7,13 +9,6 @@ terraform {
       source  = "hashicorp/google"
       version = "6.20.0"
     }
-  }
-
-  # Configuration
-  # https://developer.hashicorp.com/terraform/language/backend/gcs#prefix
-  backend "gcs" {
-    bucket = "dock-infrastructure-safehouse"
-    prefix = "terraform/environment=production"
   }
 }
 
@@ -39,15 +34,19 @@ provider "google-beta" {
   }
 }
 
+data "google_project" "current" {
+  project_id = var.project_id
+}
+
 # Search by: `gcloud services list --available | grep .googleapis.com`
 # Validate by: `gcloud services list --enabled`
 resource "google_project_service" "discovery_mesh" {
   for_each = toset([
-    "vpcaccess.googleapis.com",
-    "networksecurity.googleapis.com", # For firewall table
-    "servicenetworking.googleapis.com",
-    "compute.googleapis.com",
-    "iam.googleapis.com",
+    "vpcaccess.googleapis.com",         # Required for Serverless VPC access
+    "networksecurity.googleapis.com",   # Required for firewall/security policies
+    "servicenetworking.googleapis.com", # Required for managed services networking
+    "compute.googleapis.com",           # Required for Compute Engine VMs
+    "iam.googleapis.com",               # Required for IAM roles and service accounts
   ])
   project = var.project_id
   service = each.key
@@ -63,24 +62,4 @@ resource "google_project_service" "discovery_mesh" {
   # destroy the project, but we need the APIs available to destroy the
   # underlying resources.
   disable_on_destroy = false
-}
-
-# The numeric identifier of the project
-# See: https://registry.terraform.io/providers/hashicorp/google/latest/docs/data-sources/project
-data "google_project" "snapfire" {
-  project_id = var.project_id
-}
-
-# ---------------------------------------------------------------------------------------------
-# Outputs
-# ---------------------------------------------------------------------------------------------
-
-output "latency_app_public_ip" {
-  value       = google_compute_instance.latency_app.network_interface[0].access_config[0].nat_ip
-  description = "Public IP of the latency application VM"
-}
-
-output "latency_target_private_ip" {
-  value       = google_compute_instance.latency_target.network_interface[0].network_ip
-  description = "Private IP of the latency target VM"
 }