+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # CHANGELOG
 
-## **2025-12-26**
+## **2025-12-29**
 
 Bump project version into `0.5.1`
 
 - [x] (chore) Project layout components
 
-- [x] (dev) Development environment backed by compose services from docker-compose with
+- [x] (infrastructure) Provision infrastructure for project within GCP
 
 - [x] (docs) Updated the documentation for project (README, sad, changelog, etc.)
+
+- [x] (application) Build applications of probe instance, included: `agent`, `collector`
+
+- [x] (application) Build application for target instance, included: `exposer` with single `/_/health` endpoint